requirements.txt
fastapi
uvicorn
threads-api @ git+https://github.com/Danie1/threads-api.git